Move GetMetadata into traits folder
